+++ sdk/nodejs/policy/server.ts
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 const grpc = require("grpc");
+const analyzerproto = require("@pulumi/pulumi/proto/analyzer_pb.js");
 const analyzerrpc = require("@pulumi/pulumi/proto/analyzer_grpc_pb.js");
 const plugproto = require("@pulumi/pulumi/proto/plugin_pb.js");
 
@@ -36,6 +37,7 @@ import {
 } from "./policy";
 import {
     asGrpcError,
+    convertEnforcementLevel,
     Diagnostic,
     makeAnalyzeResponse,
     makeAnalyzerInfo,
@@ -58,6 +60,8 @@ let servingPolicyPack: string | undefined = undefined;
 
 // Regular expression for the policy pack name.
 const packNameRE = "^[a-zA-Z0-9-_.]{1,100}$";
+
+let policyPackConfig: Record<string, any> = {};
 
 /**
   * Starts the gRPC server to communicate with the Pulumi CLI client for analyzing resources.
@@ -87,6 +91,17 @@ export function serve(
             console.error(`Invalid policy name "all". "all" is a reserved name.`);
             process.exit(1);
         }
+
+        if (policy.configSchema) {
+            if (policy.configSchema.properties?.enforcementLevel) {
+                console.error(`enforcementLevel cannot be explicitly specified in properties.`);
+                process.exit(1);
+            }
+            if (policy.configSchema.required && policy.configSchema.required.includes("enforcementLevel")) {
+                console.error(`"enforcementLevel" cannot be specified in required.`);
+                process.exit(1);
+            }
+        }
     }
 
     if (servingPolicyPack) {
@@ -105,6 +120,7 @@ export function serve(
         analyzeStack: makeAnalyzeStackRpcFun(policyPackName, policyPackVersion, policyPackEnforcementLevel, policies),
         getAnalyzerInfo: makeGetAnalyzerInfoRpcFun(policyPackName, policyPackVersion, policyPackEnforcementLevel, policies),
         getPluginInfo: getPluginInfoRpc,
+        configure: configure,
     });
     const port: number = server.bind(`0.0.0.0:0`, grpc.ServerCredentials.createInsecure());
 
@@ -123,8 +139,7 @@ function makeGetAnalyzerInfoRpcFun(
 ) {
     return async function (call: any, callback: any): Promise<void> {
         try {
-            const enabledPolicies = (policies || []).filter(p => (p.enforcementLevel || policyPackEnforcementLevel) !== "disabled");
-            callback(undefined, makeAnalyzerInfo(policyPackName, policyPackVersion, policyPackEnforcementLevel, enabledPolicies));
+            callback(undefined, makeAnalyzerInfo(policyPackName, policyPackVersion, policyPackEnforcementLevel, policies));
         } catch (e) {
             callback(asGrpcError(e), undefined);
         }
@@ -136,6 +151,27 @@ async function getPluginInfoRpc(call: any, callback: any): Promise<void> {
         const resp: any = new plugproto.PluginInfo();
         resp.setVersion(version);
         callback(undefined, resp);
+    } catch (e) {
+        callback(asGrpcError(e), undefined);
+    }
+}
+
+async function configure(call: any, callback: any): Promise<void> {
+    const req = call.request;
+    try {
+        const config: Record<string, any> = {};
+        for (const [k, v] of req.getPolicyconfigMap().entries()) {
+            const props = v.getProperties().toJavaScript();
+            props.enforcementLevel = convertEnforcementLevel(v.getEnforcementlevel());
+            config[k] = props;
+        }
+        policyPackConfig = config;
+        // We need to return an new instance of `Empty` from the "google-protobuf/google/protobuf/empty_pb.js" module,
+        // but we can't use `Empty` from the module that this package depends on -- it must be from the module that
+        // @pulumi/pulumi depends on, because there is an `arg instanceof google_protobuf_empty_pb.Empty` check that
+        // will fail if `Empty` isn't from the same module. To workaround, we can simply use
+        // `AnalyzerService.configure.responseType`, which is set to `Empty` in @pulumi/pulumi.
+        callback(undefined, new analyzerrpc.AnalyzerService.configure.responseType());
     } catch (e) {
         callback(asGrpcError(e), undefined);
     }
@@ -157,7 +193,9 @@ function makeAnalyzeRpcFun(
         const ds: Diagnostic[] = [];
         try {
             for (const p of policies) {
-                const enforcementLevel = p.enforcementLevel || policyPackEnforcementLevel;
+                const enforcementLevel: EnforcementLevel =
+                    policyPackConfig[p.name]?.enforcementLevel || p.enforcementLevel || policyPackEnforcementLevel;
+
                 if (enforcementLevel === "disabled" || !isResourcePolicy(p)) {
                     continue;
                 }
@@ -175,7 +213,7 @@ function makeAnalyzeRpcFun(
                         message: violationMessage,
                         urn,
                         description: p.description,
-                        enforcementLevel: enforcementLevel,
+                        enforcementLevel,
                     });
                 };
 
@@ -208,6 +246,8 @@ function makeAnalyzeRpcFun(
                                     ? props as Unwrap<NonNullable<TArgs>>
                                     : undefined;
                             },
+
+                            getConfig: makeGetConfigFun(p.name),
                         };
                         const provider = getProviderResource(req);
                         if (provider) {
@@ -277,7 +317,9 @@ function makeAnalyzeStackRpcFun(
         const ds: Diagnostic[] = [];
         try {
             for (const p of policies) {
-                const enforcementLevel = p.enforcementLevel || policyPackEnforcementLevel;
+                const enforcementLevel: EnforcementLevel =
+                    policyPackConfig[p.name]?.enforcementLevel || p.enforcementLevel || policyPackEnforcementLevel;
+
                 if (enforcementLevel === "disabled" || !isStackPolicy(p)) {
                     continue;
                 }
@@ -295,7 +337,7 @@ function makeAnalyzeStackRpcFun(
                         message: violationMessage,
                         urn,
                         description: p.description,
-                        enforcementLevel: enforcementLevel,
+                        enforcementLevel,
                     });
                 };
 
@@ -375,6 +417,7 @@ function makeAnalyzeStackRpcFun(
 
                     const args: StackValidationArgs = {
                         resources: intermediates.map(r => r.resource),
+                        getConfig: makeGetConfigFun(p.name),
                     };
 
                     // Pass the result of the validate call to Promise.resolve.
@@ -405,6 +448,22 @@ function makeAnalyzeStackRpcFun(
 
         // Now marshal the results into a resulting diagnostics list, and invoke the callback to finish.
         callback(undefined, makeAnalyzeResponse(ds));
+    };
+}
+
+// Creates a function for retrieving the configuration for a policy.
+function makeGetConfigFun<T>(policyName: string) {
+    return function(): T {
+        // If we don't have config, or don't have config for this policy,
+        // return an empty object.
+        const c = policyPackConfig[policyName];
+        if (!c) {
+            return <T>{};
+        }
+
+        // Otherwise, return the config properties (except enforcementLevel).
+        const { enforcementLevel: ef, ...properties } = c;
+        return properties;
     };
 }