+++ tests/integration/invalid_policy/program/index.ts
@@ -1,1 +1,3 @@
 // Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
+
+// Empty program.